Check for installed directory instead of class (#44)
This change adjusts the code that guards against two versions of the
wc-api-dev plugin running at the same time. Previously the code would
check for the presence of the WC_API_Dev class, but this doesn't work
with wpcomsh composer because it runs in mu-plugins, before any other
plugins are activated.

Instead, this commit checks for the presence of the wc-api-dev plugin
installed within the conventional plugins directory. Again, this is
temporary just until all sites are updated with this new version, after
which this code check can be removed.